Fill implemented, -d option added
Fill will work with any contiguous block of pixels. This includes only
the above, below, left, right pixels and does not propagate diagonally.

Changed how the eraser works, it is not just setting the colour to the
first colour of the palette, which is by default set to 0, 0, 0, 0.
Meaning if that colour is set to a different colour, all erase
functionality will use the set colour, rather than clear.

Canvas dimensions can be set with -d width height. For example:
    $ sixel -d 16 24
will create a new canvas with a width of 16 pixels and a height of 24
pixels.